# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: Tộc Phả API
  version: v2
  description: |
    Hợp đồng API của Tộc Phả (bản chốt v2: một gia phả chung, hai vai trò Admin và User).
    File này viết tay và là **nguồn sự thật** (DECISIONS #70): frontend sinh kiểu TypeScript từ đây
    (`npm run gen:api`), backend phải khớp (test hợp đồng từ Đợt 26). Muốn đổi API thì sửa file này trước.

    Khởi tạo từ `/v3/api-docs` của backend Đợt 8, đã bỏ `/api/family/**`. Từ Đợt 39 frontend gọi backend thật cho
    mọi endpoint (đã gỡ lớp giả lập).

    Mọi API nghiệp vụ chỉ nhận tài khoản đã được duyệt (403 `ACCOUNT_NOT_APPROVED`), **trừ** các GET xem công khai
    (thành viên, người thân, cây, sự kiện, lịch, tổng quan; DECISIONS #88): khách và tài khoản chưa duyệt cũng gọi được,
    được đánh dấu `security: [{}, {bearerAuth: []}]`. Lỗi luôn là `ProblemDetail`.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: 'Cùng origin với frontend (dev: proxy /api sang :8080)'
security:
  - bearerAuth: []
...
paths:
...
  /api/calendar/lunar-month-info:
    get:
      tags:
        - calendar
      summary: Thông tin một tháng âm
      operationId: lunarMonthInfo
      security:
        - {}
        - bearerAuth: []
      description: Số ngày (29/30), ngày dương đầu và cuối của một tháng âm, kèm tháng nhuận của năm
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: month
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: leap
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LunarMonthInfoResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/calendar/convert:
    get:
      tags:
        - calendar
      summary: Đổi ngày dương và âm
      operationId: convert
      security:
        - {}
        - bearerAuth: []
      description: Đổi ngày dương sang âm (truyền solar) hoặc âm sang dương (truyền lunarYear, lunarMonth, lunarDay, leap)
      parameters:
        - name: solar
          in: query
          description: Ngày dương, dạng yyyy-MM-dd
          required: false
          schema:
            type: string
            format: date
          example: '2026-02-17'
        - name: lunarYear
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: lunarMonth
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: lunarDay
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: leap
          in: query
          description: Ngày thuộc tháng nhuận
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConvertResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /api/events:
    get:
      tags:
        - events
      summary: Danh sách sự kiện chung
      description: 'Không cần đăng nhập (khách và tài khoản chưa duyệt cũng xem được, DECISIONS #88). Xếp theo `id` tăng dần. Chỉ gồm **sự kiện chung**; giỗ và sinh nhật do `/api/calendar/*` tự sinh từ hồ sơ thành viên.'
      operationId: listEvents
      security:
        - {}
        - bearerAuth: []
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomEvent'
    post:
      tags:
        - events
      summary: Thêm sự kiện chung
      description: 'Chỉ Admin (User gửi đề xuất, Đợt 20–21). `year = null` nghĩa là lặp hằng năm theo `calendar`; có `year` thì chỉ diễn ra một lần vào đúng ngày đó.'
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomEventInput'
      responses:
        '201':
          description: Đã thêm
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomEvent'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/events/{id}:
    get:
      tags:
        - events
      summary: Chi tiết một sự kiện chung
      description: 'Lỗi: 404 `EVENT_NOT_FOUND`.'
      operationId: getEvent
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomEvent'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - events
      summary: Sửa sự kiện chung
      description: 'Chỉ Admin. Thay toàn bộ nội dung. Lỗi: 404 `EVENT_NOT_FOUND`.'
      operationId: updateEvent
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomEventInput'
      responses:
        '200':
          description: Đã sửa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomEvent'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - events
      summary: Xóa sự kiện chung
      description: 'Chỉ Admin. Lỗi: 404 `EVENT_NOT_FOUND`.'
      operationId: deleteEvent
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '204':
          description: Đã xóa
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/calendar/upcoming:
    get:
      tags:
        - calendar
      summary: Các sự kiện sắp diễn ra
      description: |
        Giỗ, sinh nhật và sự kiện chung có ngày dương từ **hôm nay** (giờ +7, `daysUntil = 0`) đến hôm nay + `days` ngày.
        Ngày cúng theo quy tắc IDEA §7 (ngày ghi đè, tháng nhuận, tháng thiếu). Xếp theo ngày dương, cùng ngày thì Giỗ, Sinh nhật, Sự kiện chung rồi theo tên.
      operationId: upcomingOccurrences
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            format: int32
            enum:
              - 7
              - 15
              - 30
              - 90
              - 365
            default: 30
        - name: type
          in: query
          description: Chỉ lấy một loại; bỏ trống là cả ba loại
          required: false
          schema:
            $ref: '#/components/schemas/EventType'
        - name: sort
          in: query
          description: '`asc` gần nhất trước (mặc định), `desc` xa nhất trước'
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CalendarOccurrence'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/calendar/month:
    get:
      tags:
        - calendar
      summary: Lịch một tháng
      description: |
        Từng ngày của một tháng kèm sự kiện của ngày đó. `mode = solar`: `year`/`month` là tháng dương. `mode = lunar`:
        là tháng âm (`leap = true` cho tháng nhuận), các ngày đi từ mùng 1 đến hết tháng âm đó, mỗi ngày kèm ngày dương.
      operationId: calendarMonth
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: month
          in: query
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 12
        - name: mode
          in: query
          required: false
          schema:
            type: string
            enum:
              - solar
              - lunar
            default: solar
        - name: leap
          in: query
          description: Chỉ có nghĩa khi `mode = lunar`
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarMonth'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/calendar/recent:
    get:
      tags:
        - calendar
      summary: Các sự kiện vừa diễn ra
      description: Các sự kiện có ngày dương **trước hôm nay** (trong vòng một năm), gần nhất trước (`daysUntil` âm). Dùng cho dashboard.
      operationId: recentOccurrences
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CalendarOccurrence'
        '400':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    EventId:
      name: id
      in: path
      required: true
      description: '`id` của sự kiện chung'
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    ValidationError:
      description: 'Dữ liệu không hợp lệ (`VALIDATION_ERROR`, `MALFORMED_REQUEST`, lỗi lịch âm...)'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Chưa đăng nhập, phiên hết hạn hoặc sai thông tin đăng nhập (`UNAUTHENTICATED`, `INVALID_CREDENTIALS`...)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Không có quyền (`FORBIDDEN`), tài khoản chưa được duyệt (`ACCOUNT_NOT_APPROVED`) hoặc bị khóa (`ACCOUNT_LOCKED`)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Không tìm thấy (`ACCOUNT_NOT_FOUND`, `MEMBER_NOT_FOUND`, `EVENT_NOT_FOUND`, `TREE_NODE_NOT_FOUND`, `RELATIVE_NOT_FOUND`, `LINK_REQUEST_NOT_FOUND`)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
    ProblemDetail:
      type: object
      description: Lỗi theo RFC 9457; `detail` là thông báo tiếng Việt hiển thị được cho người dùng.
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: Mã lỗi máy đọc được, ví dụ `ACCOUNT_NOT_APPROVED`, `MEMBER_NOT_FOUND`.
        errors:
          type: array
          description: Lỗi theo từng trường (chỉ có khi `code = VALIDATION_ERROR` hoặc lỗi lịch âm).
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
      required:
        - field
        - message
    SolarDate:
      type: object
      description: Ngày dương đầy đủ.
      properties:
        year:
          type: integer
          format: int32
        month:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        day:
          type: integer
          format: int32
          minimum: 1
          maximum: 31
      required:
        - year
        - month
        - day
...
    LunarMonthInfoResponse:
      type: object
      properties:
        year:
          type: integer
          format: int32
        month:
          type: integer
          format: int32
        leap:
          type: boolean
        days:
          type: integer
          format: int32
        firstDay:
          type: string
          format: date
        lastDay:
          type: string
          format: date
        yearLeapMonth:
          type: integer
          format: int32
    ConvertResponse:
      type: object
      properties:
        solar:
          type: string
          format: date
        lunar:
          $ref: '#/components/schemas/LunarDateResponse'
    LunarDateResponse:
      type: object
      properties:
        year:
          type: integer
          format: int32
        month:
          type: integer
          format: int32
        day:
          type: integer
          format: int32
        leap:
          type: boolean
...
    EventType:
      type: string
      description: Loại sự kiện trên lịch. Giỗ và sinh nhật tự sinh từ hồ sơ thành viên; sự kiện chung do Admin nhập.
      enum:
        - MEMORIAL
        - BIRTHDAY
        - CUSTOM
    LunarDate:
      type: object
      description: Một ngày âm đầy đủ.
      properties:
        year:
          type: integer
          format: int32
        month:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        day:
          type: integer
          format: int32
          minimum: 1
          maximum: 30
        leap:
          type: boolean
      required:
        - year
        - month
        - day
        - leap
    CustomEvent:
      type: object
      description: |
        Sự kiện chung. `calendar` là lịch người nhập dùng để tính ngày. `year = null` nghĩa là lặp hằng năm
        (âm lịch theo quy tắc IDEA §7); có `year` thì chỉ diễn ra một lần vào đúng ngày đó.
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type:
            - string
            - 'null'
        calendar:
          type: string
          enum:
            - SOLAR
            - LUNAR
        day:
          type: integer
          format: int32
          minimum: 1
          maximum: 31
        month:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        year:
          type:
            - integer
            - 'null'
          format: int32
        leap:
          type: boolean
          description: Chỉ có nghĩa khi `calendar = LUNAR`
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - calendar
        - day
        - month
        - year
        - leap
        - createdAt
        - updatedAt
    CustomEventInput:
      type: object
      description: Nội dung sự kiện chung khi thêm hoặc sửa (PUT thay toàn bộ).
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type:
            - string
            - 'null'
          maxLength: 2000
        calendar:
          type: string
          enum:
            - SOLAR
            - LUNAR
        day:
          type: integer
          format: int32
          minimum: 1
          maximum: 31
        month:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        year:
          type:
            - integer
            - 'null'
          format: int32
          description: Bỏ trống = lặp hằng năm
        leap:
          type: boolean
          default: false
          description: Chỉ có nghĩa khi `calendar = LUNAR`
      required:
        - title
        - calendar
        - day
        - month
    CalendarOccurrence:
      type: object
      description: |
        Một lần xảy ra của giỗ, sinh nhật hoặc sự kiện chung. `title` do máy chủ dựng, ví dụ "Giỗ Cụ Nguyễn Văn Sửu"
        (họ tên nguyên văn). `eventKey` ổn định giữa các lần gọi: `MEMORIAL:{memberId}:{yyyy-MM-dd}`,
        `BIRTHDAY:{memberId}:{yyyy-MM-dd}`, `CUSTOM:{eventId}:{yyyy-MM-dd}` (ngày dương của lần xảy ra).
      properties:
        eventKey:
          type: string
        type:
          $ref: '#/components/schemas/EventType'
        title:
          type: string
        description:
          type:
            - string
            - 'null'
          description: Ghi chú của sự kiện chung; `null` với giỗ và sinh nhật
        memberId:
          type:
            - integer
            - 'null'
          format: int64
          description: Người được giỗ hoặc mừng sinh nhật; `null` với sự kiện chung
        eventId:
          type:
            - integer
            - 'null'
          format: int64
          description: Chỉ có với sự kiện chung
        solar:
          $ref: '#/components/schemas/SolarDate'
        lunar:
          $ref: '#/components/schemas/LunarDate'
        daysUntil:
          type: integer
          format: int32
          description: Số ngày tính từ hôm nay (giờ +7); 0 là hôm nay, âm là đã qua
        ordinal:
          type:
            - integer
            - 'null'
          format: int32
          description: '"Giỗ lần thứ N" (MEMORIAL) hoặc "tròn N tuổi" (BIRTHDAY); `null` khi không biết năm mất/năm sinh hoặc với sự kiện chung'
      required:
        - eventKey
        - type
        - title
        - description
        - memberId
        - eventId
        - solar
        - lunar
        - daysUntil
        - ordinal
    CalendarDay:
      type: object
      properties:
        solar:
          $ref: '#/components/schemas/SolarDate'
        lunar:
          $ref: '#/components/schemas/LunarDate'
        occurrences:
          type: array
          items:
            $ref: '#/components/schemas/CalendarOccurrence'
      required:
        - solar
        - lunar
        - occurrences
    CalendarMonth:
      type: object
      properties:
        mode:
          type: string
          enum:
            - solar
            - lunar
        year:
          type: integer
          format: int32
        month:
          type: integer
          format: int32
        leap:
          type: boolean
        days:
          type: array
          description: Theo thứ tự ngày, đủ mọi ngày của tháng
          items:
            $ref: '#/components/schemas/CalendarDay'
      required:
        - mode
        - year
        - month
        - leap
        - days
...